+++ tools/matc/src/matc/MaterialCompiler.cpp
@@ -399,6 +399,7 @@ bool MaterialCompiler::run(const Config& config) {
 
     builder
         .noSamplerValidation(config.noSamplerValidation())
+        .includeEssl1(config.includeEssl1())
         .includeCallback(includer)
         .fileName(materialFilePath.getName().c_str())
         .platform(config.getPlatform())